Refactor type imports to use resorted exam, select, and split types across components and hooks
