+++ src/components/ExamSolutionSplit.tsx
@@ -5,7 +5,7 @@ import { useParams } from "next/navigation";
 import { useCallback, useEffect, useState } from "react";
 import Split from "react-split-grid";
 
-import type { Direction } from "@/types/types";
+import type { Direction } from "@/types/split.types";
 
 export default function ExamSolutionSplit() {
   const [gridTemplate, setGridTemplate] = useState<string>("1fr 18px 1fr");
