+++ src/types/select.types.ts
@@ -1,0 +1,4 @@
+export interface OptionSelect {
+  readonly value: number;
+  readonly label: number;
+}
