+++ src/types/split.types.ts
