+++ src/utils/validation.ts
@@ -1,4 +1,4 @@
-import type { OptionSelect } from "../types/interfaces";
+import type { OptionSelect } from "../types/select.types";
 
 export const isValidYear = (year: number, yearsOptions: OptionSelect[]) =>
   yearsOptions.some((option) => option.value === year);
